+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [unreleased]
+### Changed
+- Improve README and endpoint descriptions specifying that `tsv_2_json` and `csv_2_json` endpoints support only coversion for germline submission files
+
 ## [2.7.1]
 ### Fixed
 - Replace the CMD step of the Dockerfile with an ENTRYPOINT that uses `/bin/sh -c` to evaluates the shell commands and expand environment variables.